+++ Bonvino/Clases/Varietal.cs
@@ -9,13 +9,13 @@ namespace Bonvino.Clases
     public class Varietal
     {
         public Varietal() { }
-        public Varietal(TipoUva tipoUva, string descripcion, double porcentajeComposicion) { 
+        public Varietal(TipoUva tipoUva, string descripcion, float porcentajeComposicion) { 
             this.tipoUva = tipoUva;
             this.descripcion = descripcion;
             this.porcentajeComposicion = porcentajeComposicion;
         }
         public string descripcion { get; set; }
-        public double porcentajeComposicion { get; set; }
+        public float porcentajeComposicion { get; set; }
         public TipoUva tipoUva { get; set; }
     }
 }
